Create a helper file for distributed nearest stuff
